first set of refactoring
