+++ plugins/apg-ssh/src/main/groovy/com/apgsga/ssh/rpm/tasks/InstallRpm.groovy
@@ -1,17 +1,14 @@
 package com.apgsga.ssh.rpm.tasks
 
-import org.gradle.api.tasks.TaskAction
-import org.hidetake.groovy.ssh.core.Remote
-
 class InstallRpm extends AbstractRpm {
 
     public static final String INSTALL_RPM_TASK_NAME = "installRpm"
 
-    @TaskAction
-    def taskAction() {
+    @Override
+    def doRun(Object remote, Object allowAnyHosts) {
+        preConditions()
         def apgRpmDeployConfigExt = getDeployConfig()
-        Remote remotes = getRemotes()
-        project.logger.info("${apgRpmDeployConfigExt.rpmFileName} will be install on ${remotes.getProperty('host')} using ${remotes.getProperty('user')} User")
+        project.logger.info("${apgRpmDeployConfigExt.rpmFileName} will be install on ${remote.getProperty('host')} using ${remote.getProperty('user')} User")
         project.ssh.run {
             if (apgRpmDeployConfigExt.allowAnyHosts) {
                 project.logger.info("Allowing SSH Anyhosts ")
@@ -19,7 +16,7 @@ class InstallRpm extends AbstractRpm {
                     knownHosts = allowAnyHosts
                 }
             }
-            session(remotes) {
+            session(remote) {
                 executeSudo "rpm -Uvh ${apgRpmDeployConfigExt.remoteDestFolder}/${apgRpmDeployConfigExt.rpmFileName}", pty: true
                 // JHE: this probably won't stay like that, as we won't want to delete a production RPM before having archiving it. But quick fix for now for Digiflex
                 execute "rm -f ${apgRpmDeployConfigExt.remoteDestFolder}/${apgRpmDeployConfigExt.rpmFileName}", pty: true
