+++ plugins/apg-ssh/src/main/groovy/com/apgsga/ssh/zip/tasks/DeployZip.groovy
@@ -1,16 +1,14 @@
 package com.apgsga.ssh.zip.tasks
 
-import org.hidetake.groovy.ssh.core.Remote
-
 class DeployZip extends AbstractZip {
 
     public static final String DEPLOY_ZIP_TASK_NAME = "deployZip"
 
     @Override
-    def taskAction() {
+    def doRun(Object remote, Object allowAnyHosts) {
+        preConditions()
         def apgZipDeployConfigExt = getDeployConfig()
-        Remote remotes = getRemotes()
-        project.logger.info("${apgZipDeployConfigExt.zipFileName} will be deploy on ${remotes.getProperty('host')} using ${remotes.getProperty('user')} User")
+        project.logger.info("${apgZipDeployConfigExt.zipFileName} will be deploy on ${remote.getProperty('host')} using ${remote.getProperty('user')} User")
         project.ssh.run {
             if (apgZipDeployConfigExt.allowAnyHosts) {
                 project.logger.info("Allowing SSH Anyhosts ")
@@ -18,7 +16,7 @@ class DeployZip extends AbstractZip {
                     knownHosts = allowAnyHosts
                 }
             }
-            session(remotes) {
+            session(remote) {
                 put from: new File("${apgZipDeployConfigExt.zipFilePath}" + File.separator + apgZipDeployConfigExt.zipFileName), into: "${apgZipDeployConfigExt.remoteDeployDestFolder}"
                 execute("ls -la ${apgZipDeployConfigExt.remoteDeployDestFolder}") { result ->
                     println result
@@ -26,4 +24,5 @@ class DeployZip extends AbstractZip {
             }
         }
     }
+
 }
