+++ plugins/apg-ssh/src/main/groovy/com/apgsga/ssh/zip/tasks/InstallZip.groovy
@@ -1,17 +1,16 @@
 package com.apgsga.ssh.zip.tasks
 
 import com.apgsga.ssh.extensions.ApgZipDeployConfig
-import org.hidetake.groovy.ssh.core.Remote
 
 class InstallZip extends AbstractZip {
 
     public static final String INTALL_ZIP_TASK_NAME = "installZip"
 
     @Override
-    def taskAction() {
+    def doRun(Object remote, Object allowAnyHosts) {
+        preConditions()
         def apgZipDeployConfigExt = getDeployConfig()
-        Remote remotes = getRemotes()
-        project.logger.info("${apgZipDeployConfigExt.zipFileName} will be install on ${remotes.getProperty('host')} using ${remotes.getProperty('user')} User")
+        project.logger.info("${apgZipDeployConfigExt.zipFileName} will be install on ${remote.getProperty('host')} using ${remote.getProperty('user')} User")
         def unzipCmd = getUnzipCmd(apgZipDeployConfigExt)
         project.ssh.run {
             if (apgZipDeployConfigExt.allowAnyHosts) {
@@ -20,7 +19,7 @@ class InstallZip extends AbstractZip {
                     knownHosts = allowAnyHosts
                 }
             }
-            session(remotes) {
+            session(remote) {
                 executeSudo unzipCmd, pty: true
                 // JHE: it probably won't sty like that, we might not want to delete ZIP which were built for production
                 execute "rm -f ${apgZipDeployConfigExt.remoteDeployDestFolder}/${apgZipDeployConfigExt.zipFileName}", pty: true
